simplify codebase
Remove the LBT folder because the LBT algorithm still doesn't have guaranteed
convergence in LBT conditions. Those wanting it can access it through the git
history.

The ordering code was switched out to perform random selection between all
unordered children of all ordered nodes instead of performing a recursive
primary chain discovery. This is also believed to have greater fairness, as
lone disconnected miners are more likely to be included sooner.

The doublespendsim was removed as well because it is strongly believed to be
incorrect. Also, an adjustment of the max gap requirement more closely aligns
the double spend costs with traditional Bitcoin double spend attempts.